Add POSIX-compatible make description file
The make(1) utility shall derive an executable dnssec-reverb from the
dnssec-reverb.sh source file.

This is the only purpose of the Makefile; No further changes are
expected at this time. There are no GNU Make extensions, such as .PHONY
targets or task running tasks like "clean", "dist", or "install". The
aim of this is to be a highly portable, POSIX-compatible make
description file.

The file is named "Makefile" not "makefile", due to the GNU Project's
recommendation for the prior. (See
https://www.gnu.org/software/make/manual/html_node/Makefile-Names.html)

Signed-off-by: Danny Fullerton <northox@mantor.org>