Locate commands from $PATH with which(1)
which(1) is not defined by POSIX, but it makes a lot more sense than
hardcoding some paths and expecting the user to adjust a configuration
file to fit their environment. No, that's what the $PATH variable is
already for.

A `which` command first appeared in 3.0BSD.

Possibly backwards incompatible change.

Signed-off-by: Danny Fullerton <northox@mantor.org>